+++ tests/api/main.rs
@@ -4,6 +4,7 @@
 #[cfg(feature = "async")]
 mod future;
 mod listener;
+mod load_store;
 mod static_properties;
 mod tools;